requirements.txt
streamlit
openai
PyPDF2
python-docx
google-api-python-client